requirements.txt
argparse==1.4.0 ; python_version <= '3.4'
attrs==18.2.0
click-plugins==1.0.4
cligj==0.5.0
munch==2.3.2
six==1.11.0
enum34==1.1.6 ; python_version < '3.4'
